開発環境時のapiのurlを変更,dynamicルートをすべてstaticに
