+++ app/edit/layout.tsx
@@ -1,0 +1,8 @@
+import { metaDataTitle } from "@/common/title";
+import { ReactNode } from "react";
+
+export const metadata = metaDataTitle("譜面編集");
+
+export default function Page(props: { children: ReactNode[] }) {
+  return props.children;
+}
