+++ app/play/page.tsx
@@ -68,7 +68,7 @@ function InitPlay() {
     } else {
       void (async () => {
         const res = await fetch(
-          `/api/seqFile/${session.cid}/${session.lvIndex}`,
+          process.env.BACKEND_PREFIX + `/api/seqFile/${session.cid}/${session.lvIndex}`,
           { cache: "no-store" }
         );
         if (res.ok) {
