+++ tailwind.config.ts
@@ -1,7 +1,7 @@
 import type { Config } from "tailwindcss";
 
 const config: Config = {
-  content: ["./app/**/*.{js,ts,jsx,tsx,mdx}"],
+  content: ["./app/**/*.{js,ts,jsx,tsx,mdx}", "./chartFormat/**/*.{js,ts,jsx,tsx,mdx}"],
   theme: {
     screens: {
       "main-wide": "40rem", // タイトルと各タブの内容が別ページ
